// src/domain/chain-id.enum.ts
export enum ChainId {
  EthereumMain = 1,
  ArbitrumMain = 42161,
  AvalancheMain = 43114,
  BaseMain = 8453,
  BnBMain = 56,
  ZkSyncMain = 324,
  GnosisMain = 100,
  OptimismMain = 10,
  PolygonMain = 137,
  LineaMain = 59144,
  SonicMain = 146,
  UnichainMain = 130,
   
  // 추가 Layer 1
  FantomMain = 250,
  CronosMain = 25,
  CoreDaoMain = 1116,
  RootstockMain = 30,
  HaqqMain = 11235,
  ThunderCoreMain = 108,
  BitTorrentMain = 199,
  
  // 추가 Layer 2 - Ethereum
  ArbitrumNovaMain = 42170,
  ScrollMain = 534352,
  BlastMain = 81457,
  MantleMain = 5000,
  ModeMain = 34443,
  MantaPacificMain = 169,
  MetisMain = 1088,
  TaikoMain = 167000,
  ZkLinkMain = 810180,
  
  // 추가 Layer 2 - Others
  BobaMain = 288,
  KavaMain = 2222,
  CeloMain = 42220,
  FilecoinMain = 314,
  ZetaChainMain = 7000,
  
  // Emerging/Specialized
  HemiMain = 743111,
  KatanaMain = 1001012,
  HyperEvmMain = 998,
  BeraChainMain = 80084,
  PlasmaMain = 8008,
  SkaleEuropaMain = 2046399126,
  ApeChainMain = 33139,
}